fix #5: save samples as Double and implement rdb_save/load functions

+++ src/tsdb.h
@@ -11,7 +11,7 @@
 
 typedef struct Sample {
     timestamp_t timestamp;
-    int64_t data;
+    double data;
 } Sample;
 
 typedef struct Chunk
